[refactor] 안 쓰이는 Option Entity들 삭제
