+++ backend/src/main/java/com/ll/hotel/domain/hotel/hotel/entity/Hotel.java
@@ -1,6 +1,7 @@
 package com.ll.hotel.domain.hotel.hotel.entity;
 
 import com.ll.hotel.domain.hotel.hotel.type.HotelStatus;
+import com.ll.hotel.domain.hotel.option.hotelOption.entity.HotelOption;
 import com.ll.hotel.domain.hotel.room.entity.Room;
 import com.ll.hotel.domain.image.Image;
 import com.ll.hotel.domain.member.member.entity.Business;
